feat(entities): Improve Forum, Topic, User, and Comment entities
Add highs count to Forum and User entities.  Add
TopicHighsEntity to track topic highs. Improve
documentation and add validation to User DTOs. Update
related use cases and repositories for consistency.  Update
tests to reflect changes.

+++ src/main/java/br/com/soupaulodev/forumhub/modules/forum/controller/dto/ForumResponseDTO.java
@@ -28,7 +28,7 @@ public record ForumResponseDTO(UUID id,
                                UUID owner,
                                Long highs,
                                int participants,
-                               int topicCount,
+                               Long topicCount,
                                Instant createdAt,
                                Instant updatedAt) {
 }
